backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.2.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
httplib2==0.22.0
oauth2client==4.1.3
psycopg2-binary==2.9.9
pyasn1==0.6.1
pillow==10.4.0
pyasn1_modules==0.4.1
PyJWT==2.9.0
pyparsing==3.1.4
python-dotenv==1.0.1
pytz==2024.1
rsa==4.9
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1